Script to manage web_flashcards
web_flashcards is a simple Python administration script that starts, stops, or check the status of the web_flashcards html page on https://ohashisan.com/scripts/net_acronyms.html. It may take a few seconds for the page to auto-update and there may be an initial overlap in the information displayed when the page loads. I need to look at this when I have more time.